requirements.txt
autogen
firebase-admin
python-dotenv
openai
flaml[automl]